+++ lib/ratistics/math.rb
@@ -175,5 +175,38 @@ module Ratistics
       return minmax
     end
 
+    # Calculate percentile rank using the ordinal method
+    #
+    # @param [Float] percentile percentile value (0 < P < 100)
+    # @param [Integer] size the size of the data sample (N)
+    #
+    # @param [Float] the corresponding rank
+    def ordinal_rank(percentile, size)
+      # n = P / 100 * N + 1/2
+      ((percentile / 100.0 * size) + 0.5)
+    end
+
+    # Calculate percentile rank using the NIST primary method
+    #
+    # @param [Float] percentile percentile value (0 < P < 100)
+    # @param [Integer] size the size of the data sample (N)
+    #
+    # @param [Float] the corresponding rank
+    def nist_primary_rank(percentile, size)
+      # n = (P / 100) * (N + 1)
+      ((percentile / 100.0) * (size + 1))
+    end
+
+    # Calculate percentile rank using the NIST alternate method
+    #
+    # @param [Float] percentile percentile value (0 < P < 100)
+    # @param [Integer] size the size of the data sample (N)
+    #
+    # @param [Float] the corresponding rank
+    def nist_alternate_rank(percentile, size)
+      # n = ((P / 100) * (N - 1)) + 1
+      (((percentile / 100.0) * (size - 1)) + 1)
+    end
+
   end
 end